San: Rename 'guard' to 'san'
This prepares the foundation for more sanitizer-related work in the
future.
